# core: vectorize cv::normalize / cv::norm