… modules/core/src/norm.dispatch.cpp
… modules/core/src/norm.rvv1p0.hpp
… modules/core/src/norm.simd.hpp