@asmorkalov 2025-02-10
Hm.. The patch introduces significant regression for L1 and L2 on ARMv8 (Tested with Jetson Orin):
```
Norm1Arg::OCL_NormFixture::(640x480, 8UC1, NORM_INF) 	0.053 	0.010 	5.51
Norm1Arg::OCL_NormFixture::(640x480, 8UC1, NORM_L1) 	0.020 	0.067 	0.30
Norm1Arg::OCL_NormFixture::(640x480, 8UC1, NORM_L2) 	0.042 	0.071 	0.59
Norm1Arg::OCL_NormFixture::(640x480, 32FC1, NORM_INF) 	0.563 	0.042 	13.26
Norm1Arg::OCL_NormFixture::(640x480, 32FC1, NORM_L1) 	0.214 	0.142 	1.51
Norm1Arg::OCL_NormFixture::(640x480, 32FC1, NORM_L2) 	0.198 	0.121 	1.64
Norm1Arg::OCL_NormFixture::(640x480, 8UC3, NORM_INF) 	0.158 	0.031 	5.04
Norm1Arg::OCL_NormFixture::(640x480, 8UC3, NORM_L1) 	0.060 	0.200 	0.30
Norm1Arg::OCL_NormFixture::(640x480, 8UC3, NORM_L2) 	0.126 	0.213 	0.59
Norm1Arg::OCL_NormFixture::(640x480, 32FC3, NORM_INF) 	1.688 	0.138 	12.27
Norm1Arg::OCL_NormFixture::(640x480, 32FC3, NORM_L1) 	0.639 	0.428 	1.49
Norm1Arg::OCL_NormFixture::(640x480, 32FC3, NORM_L2) 	0.599 	0.361 	1.66
Norm1Arg::OCL_NormFixture::(640x480, 8UC4, NORM_INF) 	0.212 	0.043 	4.93
Norm1Arg::OCL_NormFixture::(640x480, 8UC4, NORM_L1) 	0.081 	0.267 	0.30
Norm1Arg::OCL_NormFixture::(640x480, 8UC4, NORM_L2) 	0.167 	0.284 	0.59
Norm1Arg::OCL_NormFixture::(640x480, 32FC4, NORM_INF) 	2.248 	0.193 	11.66
Norm1Arg::OCL_NormFixture::(640x480, 32FC4, NORM_L1) 	0.873 	0.589 	1.48
Norm1Arg::OCL_NormFixture::(640x480, 32FC4, NORM_L2) 	0.812 	0.506 	1.61
Norm1Arg::OCL_NormFixture::(1280x720, 8UC1, NORM_INF) 	0.159 	0.031 	5.13
Norm1Arg::OCL_NormFixture::(1280x720, 8UC1, NORM_L1) 	0.061 	0.199 	0.31
Norm1Arg::OCL_NormFixture::(1280x720, 8UC1, NORM_L2) 	0.126 	0.212 	0.59
Norm1Arg::OCL_NormFixture::(1280x720, 32FC1, NORM_INF) 	1.688 	0.138 	12.26
Norm1Arg::OCL_NormFixture::(1280x720, 32FC1, NORM_L1) 	0.650 	0.425 	1.53
Norm1Arg::OCL_NormFixture::(1280x720, 32FC1, NORM_L2) 	0.601 	0.361 	1.66
Norm1Arg::OCL_NormFixture::(1280x720, 8UC3, NORM_INF) 	0.475 	0.101 	4.71
Norm1Arg::OCL_NormFixture::(1280x720, 8UC3, NORM_L1) 	0.179 	0.593 	0.30
Norm1Arg::OCL_NormFixture::(1280x720, 8UC3, NORM_L2) 	0.377 	0.634 	0.60
Norm1Arg::OCL_NormFixture::(1280x720, 32FC3, NORM_INF) 	5.050 	0.818 	6.17
Norm1Arg::OCL_NormFixture::(1280x720, 32FC3, NORM_L1) 	1.914 	1.266 	1.51
Norm1Arg::OCL_NormFixture::(1280x720, 32FC3, NORM_L2) 	1.786 	1.075 	1.66
Norm1Arg::OCL_NormFixture::(1280x720, 8UC4, NORM_INF) 	0.655 	0.137 	4.79
Norm1Arg::OCL_NormFixture::(1280x720, 8UC4, NORM_L1) 	0.248 	0.795 	0.31
Norm1Arg::OCL_NormFixture::(1280x720, 8UC4, NORM_L2) 	0.505 	0.847 	0.60
Norm1Arg::OCL_NormFixture::(1280x720, 32FC4, NORM_INF) 	6.731 	1.164 	5.78
Norm1Arg::OCL_NormFixture::(1280x720, 32FC4, NORM_L1) 	2.539 	1.692 	1.50
Norm1Arg::OCL_NormFixture::(1280x720, 32FC4, NORM_L2) 	2.375 	1.440 	1.65
Norm1Arg::OCL_NormFixture::(1920x1080, 8UC1, NORM_INF) 	0.360 	0.074 	4.84
Norm1Arg::OCL_NormFixture::(1920x1080, 8UC1, NORM_L1) 	0.136 	0.447 	0.30
Norm1Arg::OCL_NormFixture::(1920x1080, 8UC1, NORM_L2) 	0.283 	0.475 	0.59
Norm1Arg::OCL_NormFixture::(1920x1080, 32FC1, NORM_INF) 	3.787 	0.564 	6.72
Norm1Arg::OCL_NormFixture::(1920x1080, 32FC1, NORM_L1) 	1.424 	0.949 	1.50
Norm1Arg::OCL_NormFixture::(1920x1080, 32FC1, NORM_L2) 	1.334 	0.806 	1.66
Norm1Arg::OCL_NormFixture::(1920x1080, 8UC3, NORM_INF) 	1.079 	0.343 	3.15
Norm1Arg::OCL_NormFixture::(1920x1080, 8UC3, NORM_L1) 	0.461 	1.364 	0.34
Norm1Arg::OCL_NormFixture::(1920x1080, 8UC3, NORM_L2) 	0.888 	1.494 	0.59
Norm1Arg::OCL_NormFixture::(1920x1080, 32FC3, NORM_INF) 	11.359 	2.103 	5.40
Norm1Arg::OCL_NormFixture::(1920x1080, 32FC3, NORM_L1) 	4.318 	2.836 	1.52
Norm1Arg::OCL_NormFixture::(1920x1080, 32FC3, NORM_L2) 	3.999 	2.412 	1.66
Norm1Arg::OCL_NormFixture::(1920x1080, 8UC4, NORM_INF) 	1.427 	0.533 	2.68
Norm1Arg::OCL_NormFixture::(1920x1080, 8UC4, NORM_L1) 	0.604 	1.776 	0.34
Norm1Arg::OCL_NormFixture::(1920x1080, 8UC4, NORM_L2) 	1.134 	1.898 	0.60
Norm1Arg::OCL_NormFixture::(1920x1080, 32FC4, NORM_INF) 	15.141 	2.876 	5.27
Norm1Arg::OCL_NormFixture::(1920x1080, 32FC4, NORM_L1) 	5.868 	3.791 	1.55
Norm1Arg::OCL_NormFixture::(1920x1080, 32FC4, NORM_L2) 	5.333 	3.230 	1.65 
```