@fengyuentau 2025-02-13
Regression on RISC-V for some reason:

https://github.com/opencv/opencv/actions/runs/13302059428/job/37145128708?pr=26885#step:16:715

```
[ RUN      ] Core_Array.basic_operations
/home/ci/opencv/modules/ts/src/ts.cpp:612: Failure
Failed

	failure reason: Invalid function output
	test case #-1
	seed: 00000000000c5a61
-----------------------------------
	LOG:
4: The norms are different: 801.37103666866482854/7290.6957819068729805/1996.2299161095104409 vs 801.37103666866482854/7290.69580078125/1996.2299466744807432

-----------------------------------

[  FAILED  ] Core_Array.basic_operations (24 ms)
```

Tested on my SpaceMIT MUSE Pi (K1) with GCC, it was green.