@fengyuentau 2025-02-17
@asmorkalov Hello, I have fixed performance on Orin with `gcc == 11.4.0`. See the attached file for performance.
[orin-perf.zip](https://github.com/user-attachments/files/18822169/orin-perf.zip)