@fengyuentau 2025-02-17
By the way, performance comparison with https://github.com/opencv/opencv/pull/26887 on K1.

[k1.zip](https://github.com/user-attachments/files/18823443/k1.zip)