@asmorkalov 2025-02-18
The last version is much better for ARMv8. Jetson Orin does not show regressions.
[jetson-orin-5.zip](https://github.com/user-attachments/files/18841321/jetson-orin-5.zip)