@asmorkalov 2025-02-18
Jetson tk1 (armv7) result. There are several regressions for FP32. I do not think it's critical.
[jetson-tk1-2.zip](https://github.com/user-attachments/files/18843144/jetson-tk1-2.zip)